+++ runtime_v10/src/runtime/policy.rs
@@ -394,73 +394,54 @@ pub mod policy_constants {
     pub const MAINNET_MINIMUM_CONSENSUS_POWER: i64 = 10 << 40;
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use anyhow::*;
-    use quickcheck_macros::quickcheck;
-
-    impl quickcheck::Arbitrary for Policy {
-        fn arbitrary(g: &mut quickcheck::Gen) -> Self {
-            Self {
-                max_aggregated_sectors: u32::arbitrary(g) as _,
-                min_aggregated_sectors: u32::arbitrary(g) as _,
-                max_aggregated_proof_size: u32::arbitrary(g) as _,
-                max_replica_update_proof_size: u32::arbitrary(g) as _,
-                pre_commit_sector_batch_max_size: u32::arbitrary(g) as _,
-                prove_replica_updates_max_size: u32::arbitrary(g) as _,
-                expired_pre_commit_clean_up_delay: i64::arbitrary(g),
-                wpost_proving_period: ChainEpoch::arbitrary(g),
-                wpost_challenge_window: ChainEpoch::arbitrary(g),
-                wpost_period_deadlines: u32::arbitrary(g) as _,
-                wpost_max_chain_commit_age: ChainEpoch::arbitrary(g),
-                wpost_dispute_window: ChainEpoch::arbitrary(g),
-                sectors_max: u32::arbitrary(g) as _,
-                max_partitions_per_deadline: u32::arbitrary(g) as _,
-                max_control_addresses: u32::arbitrary(g) as _,
-                max_peer_id_length: u32::arbitrary(g) as _,
-                max_multiaddr_data: u32::arbitrary(g) as _,
-                addressed_partitions_max: u32::arbitrary(g) as _,
-                declarations_max: u32::arbitrary(g) as _,
-                addressed_sectors_max: u32::arbitrary(g) as _,
-                max_pre_commit_randomness_lookback: ChainEpoch::arbitrary(g),
-                pre_commit_challenge_delay: ChainEpoch::arbitrary(g),
-                wpost_challenge_lookback: ChainEpoch::arbitrary(g),
-                fault_declaration_cutoff: ChainEpoch::arbitrary(g),
-                fault_max_age: ChainEpoch::arbitrary(g),
-                worker_key_change_delay: ChainEpoch::arbitrary(g),
-                min_sector_expiration: i64::arbitrary(g),
-                max_sector_expiration_extension: i64::arbitrary(g),
-                deal_limit_denominator: u32::arbitrary(g) as _,
-                consensus_fault_ineligibility_duration: ChainEpoch::arbitrary(g),
-                new_sectors_per_period_max: u32::arbitrary(g) as _,
-                chain_finality: ChainEpoch::arbitrary(g),
-                valid_post_proof_type: HashSet::default(),
-                valid_pre_commit_proof_type: HashSet::default(),
-                minimum_verified_allocation_size: u128::arbitrary(g).into(),
-                minimum_verified_allocation_term: i64::arbitrary(g),
-                maximum_verified_allocation_term: i64::arbitrary(g),
-                maximum_verified_allocation_expiration: i64::arbitrary(g),
-                end_of_life_claim_drop_period: ChainEpoch::arbitrary(g),
-                deal_updates_interval: i64::arbitrary(g),
-                prov_collateral_percent_supply_num: i64::arbitrary(g),
-                prov_collateral_percent_supply_denom: i64::arbitrary(g),
-                market_default_allocation_term_buffer: i64::arbitrary(g),
-                minimum_consensus_power: u128::arbitrary(g).into(),
-            }
+#[cfg(feature = "arb")]
+impl quickcheck::Arbitrary for Policy {
+    fn arbitrary(g: &mut quickcheck::Gen) -> Self {
+        Self {
+            max_aggregated_sectors: u32::arbitrary(g) as _,
+            min_aggregated_sectors: u32::arbitrary(g) as _,
+            max_aggregated_proof_size: u32::arbitrary(g) as _,
+            max_replica_update_proof_size: u32::arbitrary(g) as _,
+            pre_commit_sector_batch_max_size: u32::arbitrary(g) as _,
+            prove_replica_updates_max_size: u32::arbitrary(g) as _,
+            expired_pre_commit_clean_up_delay: i64::arbitrary(g),
+            wpost_proving_period: ChainEpoch::arbitrary(g),
+            wpost_challenge_window: ChainEpoch::arbitrary(g),
+            wpost_period_deadlines: u32::arbitrary(g) as _,
+            wpost_max_chain_commit_age: ChainEpoch::arbitrary(g),
+            wpost_dispute_window: ChainEpoch::arbitrary(g),
+            sectors_max: u32::arbitrary(g) as _,
+            max_partitions_per_deadline: u32::arbitrary(g) as _,
+            max_control_addresses: u32::arbitrary(g) as _,
+            max_peer_id_length: u32::arbitrary(g) as _,
+            max_multiaddr_data: u32::arbitrary(g) as _,
+            addressed_partitions_max: u32::arbitrary(g) as _,
+            declarations_max: u32::arbitrary(g) as _,
+            addressed_sectors_max: u32::arbitrary(g) as _,
+            max_pre_commit_randomness_lookback: ChainEpoch::arbitrary(g),
+            pre_commit_challenge_delay: ChainEpoch::arbitrary(g),
+            wpost_challenge_lookback: ChainEpoch::arbitrary(g),
+            fault_declaration_cutoff: ChainEpoch::arbitrary(g),
+            fault_max_age: ChainEpoch::arbitrary(g),
+            worker_key_change_delay: ChainEpoch::arbitrary(g),
+            min_sector_expiration: i64::arbitrary(g),
+            max_sector_expiration_extension: i64::arbitrary(g),
+            deal_limit_denominator: u32::arbitrary(g) as _,
+            consensus_fault_ineligibility_duration: ChainEpoch::arbitrary(g),
+            new_sectors_per_period_max: u32::arbitrary(g) as _,
+            chain_finality: ChainEpoch::arbitrary(g),
+            valid_post_proof_type: HashSet::default(),
+            valid_pre_commit_proof_type: HashSet::default(),
+            minimum_verified_allocation_size: u128::arbitrary(g).into(),
+            minimum_verified_allocation_term: i64::arbitrary(g),
+            maximum_verified_allocation_term: i64::arbitrary(g),
+            maximum_verified_allocation_expiration: i64::arbitrary(g),
+            end_of_life_claim_drop_period: ChainEpoch::arbitrary(g),
+            deal_updates_interval: i64::arbitrary(g),
+            prov_collateral_percent_supply_num: i64::arbitrary(g),
+            prov_collateral_percent_supply_denom: i64::arbitrary(g),
+            market_default_allocation_term_buffer: i64::arbitrary(g),
+            minimum_consensus_power: u128::arbitrary(g).into(),
         }
-    }
-
-    #[quickcheck]
-    fn policy_serde_roundtrip(policy: Policy) -> Result<()> {
-        let serialized = toml::to_string(&policy)?;
-
-        println!("serialized:\n{serialized}");
-
-        let deserialized: Policy = toml::from_str(&serialized)?;
-
-        ensure!(policy == deserialized);
-
-        Ok(())
     }
 }
